Core Functions

TrueAndErrorModels.@make_model — Macro

make_model model_type n_options n_reps

Generates a struct named model_type, constructors and a method for compute_probs based on n_options and n_reps. The struct for model_type and method for compute_probs includes auto-generated documentation.

Arguments

model_type: the name of the struct representing the model.

n_options::Vector{Int}: the number of options in each choice set, e.g., [2,3] indicates 2 options in

the first choice set and 3 in the second choice set.

n_reps: the number of times each choice set is presented. Each repetition occurs in seperate blocks with filler choices.

Example

using TrueAndErrorModels

@make_model MyCoolModel [2,2] 3

model = MyCoolModel(; p = [0.3, 0.40, 0.3, 0.0], ϵ = [0.4, 0.2, 0.2, 0.1])

Missing docstring.

Missing docstring for compute_probs. Check Documenter's build log for details.

Distributions.logpdf — Function

logpdf(
    dist::AbstractTrueErrorModel,
    data::AbstractVector{<:Integer}
)

Computes the log loglikelihood of the data for a True and Error Model.

Arguments

dist::AbstractTrueErrorModel{T}: a distribution object for a True and Error Model

data::AbstractVector{<:Integer}: a vector of response pattern frequencies.

Example

using TrueAndErrorModels

n_options = [2, 2]
n_reps = 2
@make_model TestModel n_options n_reps

model = TestModel(; p = [0.3, 0.40, 0.3, 0.0], ϵ = [0.4, 0.2, 0.2, 0.1])
data = rand(model, 100)
LL = logpdf(model, data)

Base.rand — Function

rand(
    rng::Random.AbstractRNG,
    dist::AbstractTrueErrorModel,
    n_trials::Int64
)

Generate a vector of simulated response frequencies based on the provided True and Error Model.

dist::AbstractTrueErrorModel{T}: a distribution object for a True and Error Model

n_trials: the number of simulated trials

Output

data::Vector{<:Integer}: vector of joint response frequencies with the following elements:

Example

using TrueAndErrorModels

n_options = [2, 2]
n_reps = 2
@make_model TestModel n_options n_reps

model = TestModel(; p = [0.3, 0.40, 0.3, 0.0], ϵ = [0.4, 0.2, 0.2, 0.1])
data = rand(model, 100)

Utilities

Missing docstring.

Missing docstring for get_response_labels. Check Documenter's build log for details.

TrueAndErrorModels.to_table — Function

to_table(x::Vector{<:Real})

Converts a vector of response frequencies or catetory probabilities to a 4X4 table in which the columns correspond to joint responses on the first replicate and the rows correspond to the joint responses on the second replicate. Labels S correspond to safe option and R corresponds to risky option. The position of elements corresponds to choice set.

Arguments

x::Vector{<:Real}: a vector of response frequencies or catetory probabilities with the following elements:

RR,RR

RR,RS

RR,SR

RR,SS

RS,RR

RS,RS

RS,SR

RS,SS

SR,RR

SR,RS

SR,SR

SR,SS

SS,RR

SS,RS

SS,SR

SS,SS

Returns

table::NamedArray: a 2×2 table in the following form:

4×4 Named Matrix{String}
2 ╲ 1 │      RR       RS       SR       SS
──────┼───────────────────────────────────
RR    │ "RR,RR"  "RS,RR"  "SR,RR"  "SS,RR"
RS    │ "RR,RS"  "RS,RS"  "SR,RS"  "SS,RS"
SR    │ "RR,SR"  "RS,SR"  "SR,SR"  "SS,SR"
SS    │ "RR,SS"  "RS,SS"  "SR,SS"  "SS,SS"